% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_percentile.R
\name{best_percentile}
\alias{best_percentile}
\title{Best Percentile}
\usage{
best_percentile(
  train_data,
  label,
  fitted_model,
  method,
  mean = 0,
  sd = 1,
  nboot = 200,
  snowload = TRUE,
  snowdepth_col = "snowdepth",
  snowload_col = "snowload"
)
}
\arguments{
\item{train_data}{Training data of class data.frame.}

\item{label}{Response/dependent variable name in the train_data.}

\item{fitted_model}{A fitted model from the fit_model function. A fitted
model of class "ranger" when random forest if fitted, "ksvm"
when support vector regression is fitted, and "gbm.object" when gradient
boosting machine is fitted.}

\item{method}{Machine learning model of the fitted model. This is a character class type
where "rf" - random forest, "svr" - support vector regression, and
"gbm" - gradient boosting machine.}

\item{mean}{Mean of model error. Distribution is normal.}

\item{sd}{Standard deviation of model error. Distribution is normal.}

\item{nboot}{Number of times to bootstrap the error distribution. This is an
integer type parameter. Default is 200, which creates 200 different log
normal distribution parameters.}

\item{snowload}{Logical variable indicating that the final response variable
for fitting the distribution is snowload. In this case, the initial response
variable(actual/predicted) is multiplied against snow depth. Default is
TRUE.  When FALSE, initial response variable (snowload is not computed) for
distribution fitting.}

\item{snowdepth_col}{Specify the snow depth column needed to compute the
snowload quantity. Default: "snowdepth".}

\item{snowload_col}{Specify the snowload column name needed for computing
the true parameter values. Default: "snowload".}
}
\value{
A value the represents the percentile gets the bootstrap of
predicted scale parameter close to the true parameter.
}
\description{
This function finds the best percentile out of the bootstraps of
scale parameters that minimizes the distance between the true and predicted
scale parameter. The bootstrap of log normal distribution parameters are
created by bootstrapping the error distribution from the fitted model and
adding it to the predict response variable.
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fitdistrplus]{fitdist}}
}
