% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sample_some_missing.R
\name{boot_sample_some_missing}
\alias{boot_sample_some_missing}
\title{Bootstrap fitted model error for data}
\usage{
boot_sample_some_missing(
  test_data,
  label,
  fitted_model,
  mean,
  sd,
  nboot,
  snowload,
  snowdepth_col
)
}
\arguments{
\item{test_data}{Test data of class data.frame.}

\item{label}{Response/dependent variable name in the data.}

\item{fitted_model}{A fitted model from the fit_model function. A fitted
model of class "ranger" when random forest if fitted, "ksvm"
when support vector regression is fitted, and "gbm.object" when gradient
boosting machine is fitted.}

\item{mean}{Mean of model error. Distribution is normal.}

\item{sd}{Standard deviation of model error. Distribution is normal.}

\item{nboot}{Number of times to bootstrap the error distribution. This is an
integer type parameter.}

\item{snowload}{Logical variable indicating that the final response variable
for fitting the distribution is snowload. In this case, the initial response
variable(actual/predicted) is multiplied against snow depth. Default is
TRUE.  When FALSE, initial response variable (snowload is not computed) for
distribution fitting.}

\item{snowdepth_col}{Specify the snow depth column needed to compute the
snowload quantity. Default: "snowdepth".}
}
\value{
A matrix of location and scale parameters based on nboot.
}
\description{
This is an internal function that bootstraps the error from
the fitted model and add them to the predicted response values. The new
values are then fitted to a log normal distribution.
}
