% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sample_test.R
\name{boot_sample_test}
\alias{boot_sample_test}
\title{Bootstrap fitted model error for test data}
\usage{
boot_sample_test(
  test_data,
  nboot,
  mean,
  sd,
  predictions,
  snowload,
  snowdepth_col
)
}
\arguments{
\item{test_data}{Test data of class data.frame.}

\item{nboot}{Number of times to bootstrap the error distribution. This is an
integer type parameter.}

\item{mean}{Mean of the normal distribution of fitted model errors,
Default: 0.}

\item{sd}{Standard deviation of the normal distribution of fitted model
errors, Default: 1.}

\item{predictions}{PARAM_DESCRIPTION}

\item{snowload}{Logical variable indicating that the final response variable
for fitting the distribution is snowload. In this case, the initial response
variable(actual/predicted) is multiplied against snow depth. Default is
TRUE.  When FALSE, initial response variable (snowload is not computed) for
distribution fitting.}

\item{snowdepth_col}{Specify the snow depth column needed to compute the
snowload quantity. Default: "snowdepth".}
}
\value{
A matrix of location and scale parameters based on nboot.
}
\description{
This is an internal function that bootstraps the error from
the fitted model and add them to the predicted response values. The new
values are then fitted to a log normal distribution.
}
