% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sample_train.R
\name{boot_sample_train}
\alias{boot_sample_train}
\title{Bootstrap fitted model error for train data}
\usage{
boot_sample_train(
  train_data,
  error_distr,
  nboot,
  label,
  snowload,
  snowdepth_col
)
}
\arguments{
\item{train_data}{Training data of class data.frame.}

\item{error_distr}{Fitted normal distribution of model errors.}

\item{nboot}{Number of times to bootstrap the error distribution. This is an
integer type parameter.}

\item{label}{Response/dependent variable name in the train_data.}

\item{snowload}{Logical variable indicating that the final response variable
for fitting the distribution is snowload. In this case, the initial response
variable(actual/predicted) is multiplied against snow depth. Default is
TRUE.  When FALSE, initial response variable (snowload is not computed) for
distribution fitting.}

\item{snowdepth_col}{Specify the snow depth column needed to compute the
snowload quantity. Default: "snowdepth".}
}
\value{
A matrix of location and scale parameters based on nboot.
}
\description{
This is an internal function that bootstraps the error from
the fitted model and add them to the predicted response values. The new
values are then fitted to a log normal distribution.
}
