% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumb_function.R
\name{dumb_function}
\alias{dumb_function}
\title{Dumb function}
\usage{
dumb_function(
  test_data_missing,
  test_data_not_missing,
  bounds,
  predictions,
  test_data,
  label_convert,
  multiplier,
  mean,
  sd,
  indirect_label,
  direct_label,
  distr,
  ...
)
}
\arguments{
\item{test_data_missing}{Test data with missing values for label.}

\item{test_data_not_missing}{Test data without missing values for label.}

\item{bounds}{Specify the bounds of the label of interest to ensure
a cap during the bootstrap process.}

\item{predictions}{Prediction values made for the missing label values.}

\item{test_data}{Test data of class data.frame.}

\item{label_convert}{Logical variable indicating that the final
response/label variable for fitting the distribution should be changed
from direct_label to the indirect_label. Default is FALSE, where the
direct_label is considered. If TRUE, the predicted values from the
fitted model must be multiplied by the "multiplier" to get the estimated
indirect_label.}

\item{multiplier}{Specify the multiplier column needed to compute the
indirect_label quantity. Default: "snowdepth".}

\item{mean}{Mean of model error. Distribution is normal.}

\item{sd}{Standard deviation of model error. Distribution is normal.}

\item{indirect_label}{Specify the actual indirect_label column name needed
for computing the true parameter values. Default: "snowload".}

\item{direct_label}{Response/dependent/label variable name that is
predicted by the fitted model.}

\item{distr}{A character string that represent the distribution to fit for
the label/response. Default is "lnorm" for normal distribution.}

\item{...}{Other arguments to send to the distribution function
fitdistrplus::fitdist}
}
\value{
fitted distr. object
}
\description{
Dumb function
}
