% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_true.R
\name{fit_true}
\alias{fit_true}
\title{Fit true distribution}
\usage{
fit_true(train_data, distr, direct_label, label_convert, indirect_label, ...)
}
\arguments{
\item{train_data}{Training data of class data.frame.}

\item{distr}{A character string that represent the distribution to fit for
the label/response. Default is "lnorm" for normal distribution.
See fitdistrplus::fitdist for string names for other distributions.}

\item{direct_label}{Response/dependent/label variable name that is
predicted by the fitted model.}

\item{label_convert}{Logical variable indicating that the final
response/label variable for fitting the distribution should be changed
from direct_label to the indirect_label. Default is FALSE, where the
direct_label is considered. If TRUE, the predicted values from the
fitted model must be multiplied by the "multiplier" to get the estimated
indirect_label.}

\item{indirect_label}{Specify the actual indirect_label column name needed
for computing the true parameter values. Default: "snowload".}

\item{...}{Other arguments to send to the distribution function.
fitdistrplus::fitdist}
}
\value{
Fitted distribution object
}
\description{
This internal function fits a log normal distribution to the
true response variable values.
}
