% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_distribution.R
\name{get_error_distribution}
\alias{get_error_distribution}
\title{Error distribution}
\usage{
get_error_distribution(train_data, fitted_model, method, label)
}
\arguments{
\item{train_data}{Training data of class data.frame.}

\item{fitted_model}{A fitted model from the fit_model function. A fitted
model of class "ranger" when random forest if fitted, "ksvm"
when support vector regression is fitted, and "gbm.object" when gradient
boosting machine is fitted.}

\item{method}{Machine learning model of the fitted model. This is a character class type
where "rf" - random forest, "svr" - support vector regression, and
"gbm" - gradient boosting machine.}

\item{label}{Response/dependent variable name in the train_data.}
}
\value{
Normal distribution parameters.
}
\description{
This function fits the residuals of the training data to a
normal distribution and returns the distribution parameters.
}
\seealso{
\code{\link[fitdistrplus]{fitdist}}
}
