% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_param.R
\name{predict_param}
\alias{predict_param}
\title{Predict unbiased log normal distribution parameters}
\usage{
predict_param(
  test_data,
  direct_label,
  bounds = c(-Inf, Inf),
  fitted_model,
  distr = "lnorm",
  label_convert = FALSE,
  all_missing = TRUE,
  multiplier = "snowdepth",
  indirect_label = "snowload",
  param_adjust = "sdlog",
  mean = 0,
  sd = 1,
  percentile = 0.9,
  nboot = 200,
  ...
)
}
\arguments{
\item{test_data}{Test data of class data.frame.}

\item{direct_label}{Response/dependent/label variable name that is
predicted by the fitted model.}

\item{bounds}{Specify the bounds of the label of interest to ensure
a cap during the bootstrap process.}

\item{fitted_model}{A fitted model from the fit_model function. A fitted
model of class "ranger" when random forest if fitted, "ksvm"
when support vector regression is fitted, and "gbm.object" when gradient
boosting machine is fitted.}

\item{distr}{A character string that represent the distribution to fit for
the label/response. Default is "lnorm" for normal distribution.
See fitdistrplus::fitdist for string names for other distributions.}

\item{label_convert}{Logical variable indicating that the final
response/label variable for fitting the distribution should be changed
from direct_label to the indirect_label. Default is FALSE, where the
direct_label is considered. If TRUE, the predicted values from the
fitted model must be multiplied by the "multiplier" to get the estimated
indirect_label.}

\item{all_missing}{Logical variable indicating whether the all or some
of the response variable is missing. Default: TRUE}

\item{multiplier}{Specify the multiplier column needed to compute the
indirect_label quantity. Default: "snowdepth".}

\item{indirect_label}{Specify the actual indirect_label column name needed
for computing the true parameter values. Default: "snowload".}

\item{param_adjust}{A character string that represent the distribution
parameter that needs adjustment. Default is "sdlog" from the log normal
distribution. See fitdistrplus::fitdist for string
names for other distribution parameters.}

\item{mean}{Mean of the error distribution. Distribution is normal.}

\item{sd}{Standard deviation of the error distribution. Distribution
is normal.}

\item{percentile}{Percentile of  bootstrap distribution parameter to get,
Default: 0.9.}

\item{nboot}{Number of bootstraps to run before getting the distribution
parameters, Default: 200.}

\item{...}{Other arguments to send to the distribution function
fitdistrplus::fitdist}
}
\value{
Log normal distribution parameters.
}
\description{
This function predicts an unbiased  log normal distribution
parameters of the response variable for a given set of predictor variables.
}
