% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_true.R
\name{fit_true}
\alias{fit_true}
\title{Fit true distribution}
\usage{
fit_true(train_data, label, snowload, snowload_col)
}
\arguments{
\item{train_data}{Training data of class data.frame.}

\item{label}{Response/dependent variable name in the train_data.}

\item{snowload}{Logical variable indicating that the final response variable
for fitting the distribution is snowload. In this case, the initial response
variable(actual/predicted) is multiplied against snow depth. Default is
TRUE.  When FALSE, initial response variable (snowload is not computed) for
distribution fitting.}

\item{snowload_col}{Specify the snowload column name needed for computing
the true parameter values. Default: "snowload"}
}
\value{
Fitted distribution object
}
\description{
This internal function fits a log normal distribution to the
true response variable values.
}
