% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_param.R
\name{predict_param}
\alias{predict_param}
\title{Predict log normal distribution parameters}
\usage{
predict_param(
  test_data,
  label,
  fitted_model,
  snowload = TRUE,
  all.missing = TRUE,
  snowdepth_col = snowdepth,
  snowload_col = snowload,
  mean = 0,
  sd = 1,
  percentile = 0.9,
  nboot = 200
)
}
\arguments{
\item{test_data}{Test data of class data.frame.}

\item{label}{Response/dependent variable name in the data.}

\item{fitted_model}{A fitted model from the fit_model function. A fitted
model of class "ranger" when random forest if fitted, "ksvm"
when support vector regression is fitted, and "gbm.object" when gradient
boosting machine is fitted.}

\item{snowload}{Logical variable indicating that the final response variable
for fitting the distribution is snowload. In this case, the initial response
variable(actual/predicted) is multiplied against snow depth. Default is
TRUE.  When FALSE, initial response variable (snowload is not computed) for
distribution fitting.}

\item{all.missing}{Default: TRUE}

\item{snowdepth_col}{Specify the snow depth column needed to compute the
snowload quantity. Default: "snowdepth".}

\item{snowload_col}{Specify the snowload column name needed for computing
the true parameter values. Default: "snowload".}

\item{mean}{Mean of the normal distribution of fitted model errors,
Default: 0.}

\item{sd}{Standard deviation of the normal distribution of fitted model
errors, Default: 1.}

\item{percentile}{Percentile of  bootstrap distribution parameter to get,
Default: 0.9.}

\item{nboot}{Number of bootstraps to run before getting the distribution
parameters, Default: 200.}
}
\value{
Log normal distribution parameters.
}
\description{
This function predicts the log normal distribution parameters of
of the response variable for a given set of predictor variables.
}
